+++ resources/lang/de/permissions.php
@@ -5,6 +5,8 @@ return [
     'usermgmt_manage_users' => 'Benutzerverwaltung: Benutzer erstellen, bearbeiten und löschen',
     'usermgmt_manage_roles' => 'Benutzerverwaltung: Rollen erstellen, bearbeiten und löschen',
 
+    'configure_common_settings' => 'Allgemeine Einstellungen konfigurieren',
+
     'view_changelogs' => 'Änderungshistorie betrachten',
 
     'view_logs' => 'Logeinträge betrachten',